refactor(client): extract StorageTab into standalone component
Move storage state, loaders, and UI out of SettingsPage (~1040→730 lines).
Storage API calls now only fire when the tab is opened.
